chore(search): move search from resources to kuma
